refactor: updated structure, scroll role as hook
