+++ src/features/ads/ui/ad-slot.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 
 import { GPTAdsConstants, TGPTAdsConstantsKeys } from '../consts';
-import { useAdManager } from '../hooks';
+import { useAdManager } from '../hooks/use-ad-manager';
 import { FaChevronUp } from "react-icons/fa6";
-
 
 import Image from 'next/image';
 import Link from 'next/link';
+import { useScrollAdBanner } from '../hooks/use-scroll';
 
 type TAdsSlotProps = {
   id: TGPTAdsConstantsKeys;
@@ -17,37 +17,14 @@ type TAdsSlotProps = {
 
 export function AdsSlot({ id, fixed = false }: Readonly<TAdsSlotProps>) {
   const ad = GPTAdsConstants[id];
-  const [isFixed, setIsFixed] = useState(false);
   const [bannerClose, setBannerClose] = useState(false);
-  const isFixedRef = useRef(false);
+  const { isScrolledFixed } = useScrollAdBanner(fixed)
 
-  const handleScroll = useCallback(() => {
-    if (!fixed) return;
-
-    requestAnimationFrame(() => {
-      const shouldBeFixed = window.scrollY > 250;
-
-      if (isFixedRef.current !== shouldBeFixed) {
-        isFixedRef.current = shouldBeFixed;
-        setIsFixed(shouldBeFixed);
-      }
-    });
-  }, [fixed]);
-
-  useEffect(() => {
-    if (!fixed) return;
-
-    window.addEventListener('scroll', handleScroll, { passive: true });
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, [fixed, handleScroll]);
-
-  useAdManager({
-    id,
-  });
+  useAdManager({ id });
 
   return !bannerClose && (
     <>
-      <div className={`${isFixed && 'fixed top-0 left-1/2 transform -translate-x-1/2 z-50 bg-white w-full border-gray-100 border-b-2 shadow-md shadow-gray-50'}`}>
+      <div className={`${isScrolledFixed && 'fixed top-0 left-1/2 transform -translate-x-1/2 z-50 bg-white w-full border-gray-100 border-b-2 shadow-md shadow-gray-50'}`}>
         <div className='flex flex-col items-center mt-1 md:mt-0 md:items-center'>
           <span className='font-primary font-bold text-[6px] uppercase mt-4 text-[#9E9E9E]'>
             publicidade
@@ -74,7 +51,7 @@ export function AdsSlot({ id, fixed = false }: Readonly<TAdsSlotProps>) {
           </div>
         </div>
 
-        {isFixed &&
+        {isScrolledFixed &&
           <>
             <FaChevronUp
               className='cursor-pointer z-10 ml-0 -mb-7 mt-1 text-2xl font-bold border-gray-100 border-2 border-t-0 py-1 px-4 h-7 w-20 bg-white rounded-b-[7px] shadow-md shadow-gray-120 hover:bg-gray-100'
@@ -83,7 +60,7 @@ export function AdsSlot({ id, fixed = false }: Readonly<TAdsSlotProps>) {
           </>
         }
       </div>
-      {isFixed && <div className="mb-40"></div>}
+      {isScrolledFixed && <div className="mb-40"></div>}
     </>
   );
 }
